FIX conflixt
Former-commit-id: b5c5e786d87550930f672e7575fccf9b721285c3

+++ Back-End/ssachedule/src/main/java/com/ssljjong/ssachedule/repository/NoticeRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface NoticeRepository extends JpaRepository<Notice, Long> {
+public interface NoticeRepository extends JpaRepository<Notice, Integer> {
 
     @Query("select n from Notice n join fetch Channel c" +
             " join c.team t join t.teamUsers tu join tu.user u where u.id = :userId")
