+++ Back-End/ssachedule/src/main/java/com/ssljjong/ssachedule/repository/TrackRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface TrackRepository extends JpaRepository<Track, Long> {
+public interface TrackRepository extends JpaRepository<Track, Integer> {
 
     /**
      * * find Track By Track_Name and Gi
